// models/User.ts
// Auto-generated from C# class User. Do not edit manually.
import type { Bill } from './Bill';
import type { Comment } from './Comment';
import type { Daily } from './Daily';
import type { DailyHistory } from './DailyHistory';
import type { Diet } from './Diet';
import type { EmployeeRegisterDaily } from './EmployeeRegisterDaily';
import type { EmployeeUser } from './EmployeeUser';
import type { FitUser } from './FitUser';
import type { Gym } from './Gym';
import type { Like } from './Like';
import type { LogChange } from './LogChange';
import type { LogError } from './LogError';
import type { LogLogin } from './LogLogin';
import type { LogUninstall } from './LogUninstall';
import type { Notification } from './Notification';
import type { NotificationOption } from './NotificationOption';
import type { PaymentAttempt } from './PaymentAttempt';
import type { Permission } from './Permission';
import type { PhysicalAssessment } from './PhysicalAssessment';
import type { Plan } from './Plan';
import type { Post } from './Post';
import type { RoutineAssigned } from './RoutineAssigned';
import type { RoutineTemplate } from './RoutineTemplate';
import type { Schedule } from './Schedule';
import type { UserType } from './UserType';

export interface User {
  Id: string;
  Email: string;
  Password: string;
  IdEps: string | null;
  Name: string | null;
  LastName: string | null;
  UserName: string | null;
  IdGender: string | null;
  BirthDate: string | null;
  ProfileImageUrl: string | null;
  DocumentTypeId: string | null;
  Phone: string | null;
  CountryId: string | null;
  Address: string | null;
  CityId: string | null;
  RegionId: string | null;
  Rh: string | null;
  EmergencyName: string | null;
  EmergencyPhone: string | null;
  PhysicalExceptions: string | null;
  PhysicalExceptionsNotes: string | null;
  CreatedAt: string | null;
  UpdatedAt: string | null;
  DeletedAt: string | null;
  Ip: string | null;
  IsActive: boolean | null;
  UserTypeId: string | null;
  PlanId: string | null;
  UserFitUserUserId: string | null;
  UserDietUserId: string | null;
  EmployeeRegisterDailyUserUserId: string | null;
  ScheduleUserUserId: string | null;
  UserEmployeeUserUserId: string | null;
  GymId: string | null;
  RegistrationCompleted: boolean;
  BillUserSellers: Bill[];
  BillUsers: Bill[];
  Dailies: Daily[];
  DailyHistories: DailyHistory[];
  Diets: Diet[];
  EmployeeRegisterDailyUserUser: EmployeeRegisterDaily | null;
  Gym: Gym | null;
  LogChanges: LogChange[];
  LogErrors: LogError[];
  LogLogins: LogLogin[];
  LogUninstalls: LogUninstall[];
  NotificationOptions: NotificationOption[];
  Notifications: Notification[];
  Permissions: Permission[];
  PhysicalAssessments: PhysicalAssessment[];
  Plan: Plan | null;
  RoutineAssigneds: RoutineAssigned[];
  RoutineTemplates: RoutineTemplate[];
  ScheduleUserUser: Schedule | null;
  UserDietUser: Diet | null;
  UserEmployeeUser: EmployeeUser | null;
  UserFitUser: FitUser | null;
  UserType: UserType | null;
  PaymentAttempts: PaymentAttempt[];
  Posts: Post[];
  Comments: Comment[];
  Likes: Like[];
}